backend/requirements.txt
blinker==1.9.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
colorama==0.4.6
decorator==4.4.2
et_xmlfile==2.0.0
filelock==3.16.1
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-SQLAlchemy==3.1.1
fsspec==2024.10.0
greenlet==3.1.1
idna==3.10
imageio==2.36.0
imageio-ffmpeg==0.5.1
itsdangerous==2.2.0
Jinja2==3.1.4
llvmlite==0.43.0
MarkupSafe==3.0.2
more-itertools==10.5.0
moviepy==1.0.3
mpmath==1.3.0
networkx==3.4.2
numba==0.60.0
numpy==2.0.2
openai-whisper==20240930
openpyxl==3.1.5
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pydub==0.25.1
PyJWT==2.10.0
python-dotenv==1.0.1
pyvips==2.2.3
regex==2024.11.6
requests==2.32.3
scipy==1.14.1
six==1.16.0
SQLAlchemy==2.0.36
sympy==1.13.1
tiktoken==0.8.0
torch==2.5.1
tqdm==4.67.0
typing_extensions==4.12.2
urllib3==2.2.3
Werkzeug==3.1.3
whisper==1.1.10
Flask-SQLAlchemy
Flask-JWT-Extended
PyJWT